+++ SteamWebRequest/SteamApiClient/Models/Dota/Cosmetics/Rarities.cs
@@ -5,8 +5,8 @@ namespace SteamApi.Models.Dota
     public class DotaCosmeticRarity
     {
         public string Name { get; set; }
-        public byte Id { get; set; }
-        public byte Order { get; set; }
+        public uint Id { get; set; }
+        public uint Order { get; set; }
         public string Color { get; set; }
 
         [JsonProperty("localized_name")]
