+++ SteamWebRequest/SteamApiClient/Models/Dota/GarbageParents/LiveLeagueMatchContens.cs
@@ -4,7 +4,7 @@ namespace SteamApi.Models.Dota
 {
     internal class LiveLeagueMatchContens
     {
-        public ushort Status { get; set; }
+        public uint Status { get; set; }
         public IReadOnlyCollection<LiveLeagueMatch> Games { get; set; }
     }
 }
