+++ SteamWebRequest/SteamApiClient/Models/Dota/Match/RealTimeMatchPlayer.cs
@@ -8,32 +8,32 @@ namespace SteamApi.Models.Dota
         public uint AccountId { get; set; }
         public uint PlayerId { get; set; }
         public string Name { get; set; }
-        public byte Team { get; set; }
-        public ushort HeroId { get; set; }
-        public byte Level { get; set; }
+        public uint Team { get; set; }
+        public uint HeroId { get; set; }
+        public uint Level { get; set; }
 
         [JsonProperty("kill_count")]
-        public byte Kills { get; set; }
+        public uint Kills { get; set; }
 
         [JsonProperty("assists_count")]
-        public byte Assists { get; set; }
+        public uint Assists { get; set; }
 
         [JsonProperty("death_count")]
-        public byte Deaths { get; set; }
+        public uint Deaths { get; set; }
 
         [JsonProperty("denies_count")]
-        public ushort Denies { get; set; }
+        public uint Denies { get; set; }
 
         [JsonProperty("lh_count")]
-        public ushort LastHits { get; set; }
+        public uint LastHits { get; set; }
 
-        public ushort Gold { get; set; }
+        public uint Gold { get; set; }
 
         public double X { get; set; }
         public double Y { get; set; }
 
         [JsonProperty("net_worth")]
-        public ushort NetWorth { get; set; }
+        public uint NetWorth { get; set; }
 
         public IReadOnlyCollection<ushort> Abilities { get; set; }
         public IReadOnlyCollection<ushort> Items { get; set; }
