+++ SteamWebRequest/SteamApiClient/Models/Dota/Shared/Pick.cs
@@ -5,6 +5,6 @@ namespace SteamApi.Models.Dota
     public class Pick
     {
         [JsonProperty("hero_id")]
-        public ushort HeroId { get; set; }
+        public uint HeroId { get; set; }
     }
 }
